Updates unit tests
